improve grammar print function
