+++ transformers_cfg/recognizer.py
@@ -1,245 +1,145 @@
 import copy
 import logging
-import sys
-from abc import ABC
 from functools import lru_cache
-from typing import Dict, List
+from typing import List, Tuple, Dict
 
-import torch
+from transformers_cfg.parser import (
+    END_OF_RULE_MARKER,
+    END_OF_ALTERNATE_MARKER,
+    parse_ebnf,
+    REF_RULE_MARKER,
+)
 
-from transformers_cfg.parser import parse_ebnf, print_grammar, END_OF_RULE_MARKER
-from transformers_cfg.grammar import GrammarAutomaton
-from .vocab_struct import LEAF, TokenTrie
 
-logger = logging.getLogger(__name__)
+class GrammarRecognizer:
+    def __init__(self, grammar_encoding: List[int], start_rule_id: int):
+        # strictly speaking, we don't need to copy grammar_encoding because we don't modify it
+        # but we do it anyway to be safe
+        # in case where the grammar is very large, we can consider not copying it
+        self.grammar_encoding = copy.deepcopy(grammar_encoding)
+        self.rule_offsets: Dict[int, int] = self.init_rules(start_rule_id)
+        # each stack is a list of indices into grammar_encoding
+        # each index points to a rule's
+        self.stacks: List[List[int]] = self.init_stack(start_rule_id)
 
+    def init_rules(self, start_rule_id: int) -> Dict[int, int]:
+        _rule_offset = 0
+        rule_offsets = {}
+        # Build `rules` as a dictionary of rule IDs to their positions in `grammar_src`
+        while self.grammar_encoding[_rule_offset] != 0xFFFF:
+            rule_id = self.grammar_encoding[_rule_offset]
+            # store the offset idx
+            rule_offsets[rule_id] = _rule_offset
 
-class AbstractGrammarConstraint(ABC):
-    def __init__(self, grammar_str, tokenizer, start_rule_name="root"):
-        state = parse_ebnf(grammar_str)
-        grammar_encoding = state.grammar_encoding
-        self.start_rule_id = state.symbol_table.get(start_rule_name)
+            # Skip rule ID
+            # _rule_offset += 1
+            simple_rhs_offset = _rule_offset + 1
 
-        self.eos_token_id = tokenizer.eos_token_id
-        self.token_trie = TokenTrie(tokenizer)
-        self.tokenizer = tokenizer
-        self.grammar = GrammarAutomaton(grammar_encoding, self.start_rule_id)
-
-    def _consume_token_id(self, token_id: int, stacks: List[List[int]]):
-        if token_id == self.eos_token_id:
-            if stacks and all(len(stack) != 0 for stack in stacks):
-                raise Exception(
-                    f"At least one of the stack should be empty when EOS is reached. However, "
-                    f"the stacks are {stacks}"
+            # Skip rule alternates
+            while self.grammar_encoding[simple_rhs_offset] != END_OF_RULE_MARKER:
+                simple_rhs_offset = (
+                    simple_rhs_offset + 1 + self.grammar_encoding[simple_rhs_offset]
                 )
-            return []
 
-        for byte in self.token_trie.id2str(token_id):
-            stacks = self.grammar._consume_char(byte, stacks)
-            # check updated stacks
-            # TODO, I commented this out because it will fail when the stack is empty
-            # empty stack means the end of the grammar
-            # assert stacks != []
+            # Skip 0 denoting end of rule
+            # _rule_offset += 1
+            _rule_offset = simple_rhs_offset + 1
 
+        retrieved_start_rule_id = self.grammar_encoding[rule_offsets[start_rule_id]]
+        assert retrieved_start_rule_id == start_rule_id
+
+        return rule_offsets
+
+    def init_stack(self, start_rule_id: int) -> List[List[int]]:
+
+        stacks = []
+        # Loop over alternates of start rule to build initial stacks
+        sub_rhs_offset = self.rule_offsets[start_rule_id] + 1
+        while self.grammar_encoding[sub_rhs_offset]:
+            stack: List[int] = []
+            # If alternate is nonempty, add to stack
+            element_offset = sub_rhs_offset + 1
+            if self.grammar_encoding[element_offset] != END_OF_ALTERNATE_MARKER:
+                stack.append(element_offset)
+            stacks.extend(self.advance_stack(tuple(stack)))
+            sub_rhs_offset += 1 + self.grammar_encoding[sub_rhs_offset]
         return stacks
 
-    def advance_token_ids(self, *args, **kwargs):
-        """Process a list of tokens according to the grammar rules."""
-        raise NotImplementedError
-
-    def batch_filter_vocab(self, batch_stacks, device) -> torch.Tensor:
-        batch_acceptance = []
-        for stacks in batch_stacks:
-            batch_acceptance.append(self.filter_vocab(stacks, device))
-        return torch.stack(batch_acceptance)
-
-    def filter_vocab(self, stacks, device) -> torch.Tensor:
-        if not stacks:  # Check if stacks is empty
-            # Handle the empty case: for example, return a tensor of False
-            # The size of the tensor should match the size of your vocabulary
-            vocab_size = len(self.token_trie)
-            logger.debug(f"Empty stack, sum of acceptance: {0}")
-            return torch.zeros(vocab_size, dtype=torch.bool, device=device)
-
-        acceptance_matrix = torch.cat(
-            [self.token_acceptance_for_stack(tuple(stack), device) for stack in stacks]
-        )
-        # Merge stacks: any True => True
-        acceptance = acceptance_matrix.reshape(len(stacks), -1).any(dim=0)
-        logger.debug(f"sum of acceptance: {acceptance.sum()}")
-        return acceptance
-
-    # For each sub-rule in the grammar, cache whether each byte is accepted.
-    @lru_cache(maxsize=None)
-    def char_acceptance_at_rule_pos(self, rule_offset):
-        # every time this function is called, the result is cached
-        # next time when the same pos is called, the result is returned directly
-        # Here the pos corresponds to the literal or char range rule
-        # it doesn't handle the rule reference
-        acceptance = [False] * 256
-        num_chars = self.grammar.grammar_encoding[rule_offset]
-        rule_offset += 1
-        for i in range(0, num_chars, 2):
-            start = self.grammar.grammar_encoding[rule_offset + i]
-            end = self.grammar.grammar_encoding[rule_offset + i + 1]
-            for j in range(start, end + 1):
-                acceptance[j] = True
-        return acceptance
-
-    # Probably this should be configurable. If the grammar has an exceedingly
-    # large number of states, the correct setting is a tradeoff between GPU
-    # RAM usage and recomputation time.
-    #
-    # The main variable that pushes usage up here is number of states in the
-    # grammar.
     @lru_cache(maxsize=32768)
-    def token_acceptance_for_stack(self, stack, device):
-        stack = list(stack)  # needs to come in as a tuple for lru_cache
-
-        accepts = [False] * len(self.token_trie)
-        accepts[self.eos_token_id] = len(stack) == 0
+    def advance_stack(self, stack: Tuple[int]) -> List[List[int]]:
+        stack = list(stack)
         if len(stack) == 0:
-            logger.debug("empty stack")
+            return [stack]
 
-        def traverse_trie(trie, stacks):
-            for byte, next_trie in trie.items():
-                if byte == LEAF:
-                    token_id = next_trie
-                    if token_id != self.eos_token_id:
-                        # if the stacks is not empty, it means we can still continue to parse
-                        # so we should accept the token
-                        accepts[token_id] = bool(stacks)
-                    continue
+        # we get the last element of the stack, which is the element we are currently processing
+        cur_element_offset = stack[-1]
 
-                new_stacks = []
-                for stk in stacks:
-                    if not stk:
-                        continue
-
-                    next_rule_pos = stk[-1]
-                    num_chars = self.grammar.grammar_encoding[next_rule_pos]
-
-                    if not self.char_acceptance_at_rule_pos(next_rule_pos)[byte]:
-                        # if the current byte is not accepted by the current rule, we need to try next rule
-                        continue
-
-                    next_rule_pos += num_chars + 1
-                    new_stack = stk[:-1]
-                    if self.grammar.grammar_encoding[next_rule_pos]:
-                        new_stack.append(next_rule_pos)
-                    new_stacks.extend(self.grammar.advance_stack(tuple(new_stack)))
-
-                if new_stacks:
-                    traverse_trie(next_trie, new_stacks)
-
-        traverse_trie(self.token_trie.trie, [stack])
-
-        x = torch.tensor(accepts, dtype=torch.bool, device=device)
-        return x
-
-
-class IncrementalGrammarConstraint(AbstractGrammarConstraint):
-    def __init__(self, grammar_str, start_rule_name, tokenizer):
-        super().__init__(grammar_str, tokenizer, start_rule_name)
-        self.last_size = None
-
-        # if self.last_size is not set (which would be the case when processing the first token).
-        # In this case, do nothing.
-
-    def advance_token_ids(self, input_ids, batch_stacks, parse_start_index=None):
-
-        if self.last_size is None:
-            prefix_to_parse = [
-                single_input_ids[parse_start_index:]
-                if parse_start_index is not None
-                else []
-                for single_input_ids in input_ids
-            ]
-
-            # self.grammar_acceptor.accept_token_ids(prefix_to_parse, self.stacks)
-            batch_stacks = [
-                self._consume_token_ids(prefix, stack)
-                for prefix, stack in zip(prefix_to_parse, batch_stacks)
-            ]
-            #  if the length of the current input IDs (input_ids[0]) is exactly one more than self.last_size.
-            #  This is expected in a scenario where inputs are processed incrementally, one token at a time.
-        elif len(input_ids[0]) == self.last_size + 1:
-            batch_stacks = [
-                self._consume_token_id(single_input_ids[-1], stack)
-                for single_input_ids, stack in zip(input_ids, batch_stacks)
-            ]
-            #  ensure that the input size is consistent with the expected incremental processing
-            #  (i.e., one token at a time).
+        # if the element is a terminal, we don't need to advance the stack
+        if self.grammar_encoding[cur_element_offset] != REF_RULE_MARKER:
+            return [stack]
+        # the remaining case is that the element is a non-terminal, i.e. a reference to another rule
         else:
-            # here we check if the input_ids are one token longer than the last time we processed
-            # but we don't check if input_ids are actually valid.
-            # Imagine a scenario where we generate 10 tokens, then we replace the 10 generated tokens with 10 new tokens.
-            # In this case, the input_ids will be consistent with the last_size, but the input_ids are not valid.
-            # However, should we really check if the input_ids are valid here?
-            # If we do, then we need to reparse the whole input_ids at each call, which is not efficient.
-            # Maybe we should just trust the user to provide valid input_ids?
-            # The conclusion is that, we assume the input_ids are valid, and our generation will be correct.
-            # If the input_ids are not valid, then the generation result will be wrong and we don't take responsibility for that.
-            raise RuntimeError(
-                "Input ID's length is inconsistent with the current state of "
-                "the GrammarConstrainedLogitsProcessor. If you want to process "
-                "another input sequence, please instantiate a new "
-                "GrammarConstrainedLogitsProcessor."
-            )
-        self.last_size = len(input_ids[0])
+            ref_rule_id = self.grammar_encoding[cur_element_offset + 1]
+            # find the offset of the referenced rule
+            ref_subrule_offset = self.rule_offsets[ref_rule_id] + 1
+            new_stacks: List[List[int]] = []
+            # Loop over alternates of referenced rule to build new stacks
+            while self.grammar_encoding[ref_subrule_offset] != END_OF_RULE_MARKER:
+                # copy the original stack without the last element
+                new_stack = stack[:-1]
+                # if the rule ref is followed by another element, we add it to the stack
+                next_element_offset = cur_element_offset + 2
+                if (
+                    self.grammar_encoding[next_element_offset]
+                    != END_OF_ALTERNATE_MARKER
+                ):
+                    new_stack.append(next_element_offset)
 
-        return batch_stacks
+                # if the referenced rule is not empty, we add its element offset to the stack
+                ref_element_offset = ref_subrule_offset + 1
+                if self.grammar_encoding[ref_element_offset] != END_OF_ALTERNATE_MARKER:
+                    new_stack.append(ref_element_offset)
 
-    def _consume_token_ids(
-        self, token_ids: List[int], stacks: List[List[int]], as_string=True
-    ):
-        if as_string:
-            string = self.tokenizer.decode(token_ids)
-            stacks = self.grammar._consume_string(string, stacks)
-        else:
-            for token_id in token_ids:
-                stacks = self._consume_token_id(token_id, stacks)
-        return stacks
+                new_stacks.extend(self.advance_stack(tuple(new_stack)))
+                ref_subrule_offset += self.grammar_encoding[ref_subrule_offset] + 1
 
+            return new_stacks
 
-class VanillaGrammarConstraint(AbstractGrammarConstraint):
-    # TODO, this class may have been broken because of the recent refactoring, in particular, the stacks is bound to the
-    # parser class
-    def __init__(self, grammar_str, start_rule_name, tokenizer):
-        super().__init__(grammar_str, tokenizer, start_rule_name)
-        self.offset = None
+    def _consume_char(self, byte, stacks):
+        new_stacks = []
+        for stack in stacks:
+            # stack is empty
+            if not stack:
+                continue
 
-    def advance_token_ids(self, input_ids, batch_stacks, parse_start_index=None):
-        # By design, the batch_stacks should be empty at the beginning, thus it doesn't matter what we pass in.
-        if self.offset is None:
-            self.offset = (
-                len(input_ids[0]) if parse_start_index is None else parse_start_index
-            )
+            element_offset = stack[-1]
+            size = self.grammar_encoding[element_offset]
 
-        # TODO: here may be broken, the deepcopy may not work
-        batch_stacks_from_scratch = [
-            copy.deepcopy(self.grammar.stacks) for _ in range(len(input_ids))
-        ]
+            # to make idx point to the range_start of the first range
+            element_offset += 1
+            found = False
+            for i in range(0, size, 2):
+                if (
+                    self.grammar_encoding[element_offset + i] <= byte
+                    and byte <= self.grammar_encoding[element_offset + i + 1]
+                ):
+                    found = True
+                    break
+            if not found:
+                continue
 
-        prefix_to_consume = [
-            single_input_ids[self.offset :] for single_input_ids in input_ids
-        ]
-        # self.grammar_acceptor.accept_token_ids(prefix_to_consume, self.stacks)
-        advanced_batch_stacks = [
-            self._consume_token_ids(prefix, stack, as_string=False)
-            for prefix, stack in zip(prefix_to_consume, batch_stacks_from_scratch)
-        ]
-        return advanced_batch_stacks
+            element_offset += size
+            new_stack = stack[:-1]
+            if self.grammar_encoding[element_offset]:
+                new_stack.append(element_offset)
+            new_stacks.extend(self.advance_stack(tuple(new_stack)))
 
-    def _consume_token_ids(
-        self, token_ids: List[int], stacks: List[List[int]], as_string=True
-    ):
-        if as_string:
-            string = self.tokenizer.decode(token_ids)
-            stacks = self.grammar._consume_string(string, stacks)
-        else:
-            for token_id in token_ids:
-                stacks = self._consume_token_id(token_id, stacks)
+        return new_stacks
+
+    def _consume_string(self, string: str, stacks: List[List[int]]):
+        _bytes = bytes(string, "utf-8")
+        for byte in _bytes:
+            stacks = self._consume_char(byte, stacks)
         return stacks
 
 
@@ -248,12 +148,20 @@ if __name__ == "__main__":
     logging.basicConfig(level=logging.DEBUG)
 
     try:
-        with open("examples/grammars/json.ebnf", "r") as file:
+        with open("examples/grammars/debug/plus.ebnf", "r") as file:
             input_text = file.read()
-        state = parse_ebnf(input_text)
-        print_grammar(sys.stdout, state)
-        print(f"symbol_ids: \n{state.symbol_table}")
+        parsed_grammar = parse_ebnf(input_text)
+        parsed_grammar.print()
+        print(f"symbol_ids: \n{parsed_grammar.symbol_table}")
 
+        start_rule_id = parsed_grammar.symbol_table["root"]
+        json_automaton = GrammarRecognizer(
+            parsed_grammar.grammar_encoding, start_rule_id
+        )
+        print(f"rule_offsets: \n{json_automaton.rule_offsets}")
+        print(f"stacks: \n{json_automaton.stacks}")
+        out_stacks = json_automaton._consume_char(ord("1"), json_automaton.stacks)
+        print(f"out_stacks: \n{out_stacks}")
     except FileNotFoundError:
         print("Error: File 'grammar.ebnf' not found.")
     except IOError as e:
